requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aniso8601=8.0.0=py_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
certifi=2020.6.20=pyhd3eb1b0_3
click=7.1.2=py_0
flask=1.1.2=py_0
flask-cors=3.0.9=py_0
flask-restful=0.3.8=py_0
flask-sqlalchemy=2.4.4=py_0
freetype=2.10.4=ha233b18_0
intel-openmp=2019.4=233
itsdangerous=1.1.0=py37_0
jinja2=2.11.2=py_0
jpeg=9b=he5867d9_2
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
lz4-c=1.9.2=h79c402e_3
markupsafe=1.1.1=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
ninja=1.10.1=py37h879752b_0
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
olefile=0.46=py37_0
openssl=1.1.1h=haf1e3a3_0
pillow=8.0.1=py37h5270095_0
pip=20.2.4=py37hecd8cb5_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
pytorch=1.7.0=py3.7_0
pytz=2020.1=py_0
readline=8.0=h1de35cc_0
setuptools=50.3.1=py37hecd8cb5_1
six=1.15.0=py_0
sqlalchemy=1.3.20=py37h9ed2024_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.8.1=py37_cpu
typing_extensions=3.7.4.3=py_0
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
